requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
pillow>=10.0.0
tqdm>=4.65.0
tensorboard>=2.13.0
plotly>=5.15.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0
albumentations>=1.3.0
torchsummary>=1.5.1
